Use kotlin to show adjust(radius)

+++ app/src/main/java/com/qiaoqiao/app/PrefsKeys.java
@@ -18,5 +18,5 @@ public interface PrefsKeys {
 	float DEFAULT_CONFIDENCE_IMAGE = 0.5f;
 
 	String KEY_GEOSEARCH_RADIUS = "geosearchRadius";
-	long DEFAULT_GEOSEARCH_RADIUS = 500L;
+	int DEFAULT_GEOSEARCH_RADIUS = 500L;
 }
